requirements.txt
aerich==0.9.0
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttpx==0.0.12
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-lru==2.0.5
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
colorlog==6.9.0
dictdiffer==0.9.0
distro==1.9.0
dotenv==0.9.9
frozendict==2.4.6
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
jiter==0.10.0
lazyops==0.4.0
loguru==0.7.3
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.4.4
openai==1.89.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
pypika-tortoise==0.6.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
tiktoken==0.9.0
tomlkit==0.13.3
tortoise-orm==0.25.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
win32_setctime==1.2.0
yarl==1.20.1
